pyproject.toml
[tool.poetry]
name = "transportik-lite"
version = "0.0.1"
description = ""
authors = ["Андрей Романов <romanov_andrey_vl@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
markdown = "^3.1"
django-filter = "^2.2"
django-split-settings = "^1.0"
psycopg2 = "^2.8"
gunicorn = "^19.9"
layered-yaml-attrdict-config = "^18.12"
djangorestframework-jwt = "^1.11"
bcrypt = "^3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^3.6"
pytest-cov = "^2.8"
pytest-randomly = "^3.1"
pytest-testmon = "^0.9.19"
pytest-timeout = "^1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
